replace renderSvg helper in global.js
